% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calcs}
\alias{calcs}
\title{calcs.}
\usage{
calcs(model, x)
}
\arguments{
\item{model}{The fit model}

\item{x}{Vector of xbar values}
}
\value{
Vector of model-based stdev values
}
\description{
Calculate estimated standard deviation using the fit model.
}

