% Generated by roxygen2 (4.0.2): do not edit by hand
\name{data_ok}
\alias{data_ok}
\title{data_ok.}
\usage{
data_ok(D, nA, nB)
}
\arguments{
\item{D}{A data.frame containing the entire contents of the input Excel spreadsheet: protein
names and spectral count data.}

\item{nA}{The number of replicates in the first condition.}

\item{nB}{The number of replicates in the second condition.}
}
\value{
A boolean (TRUE) if checks are passed successfully, FALSE otherwise.
}
\description{
Ensures that the input file has as many columns as claimed, and at least one row.
Also checks if the spectral-count data contains only finite positive values.
}

