% Generated by roxygen2 (4.0.2): do not edit by hand
\name{diff_exp_table}
\alias{diff_exp_table}
\title{diff_exp_table.}
\usage{
diff_exp_table(stn_pval, Prot, num_digits, outfile = "")
}
\arguments{
\item{stn_pval}{List of model-based STN and p-values, as generated by \code{\link{calc_stn_pval}}}

\item{Prot}{Vector of protein names}

\item{num_digits}{Number of digits for rounding numbers}

\item{outfile}{Name of xlsx file to which the table will be written}
}
\value{
If \code{outfile} is a blank string, the table will be returned as a data.frame. Otherwise
	it will be written to the specified xlsx file.
}
\description{
Generate table of results from differential expression test.
}

