% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fit_model}
\alias{fit_model}
\title{fit_model.}
\usage{
fit_model(A, B, fit_type)
}
\arguments{
\item{A}{Replicates in the first condition.}

\item{B}{Replicates in the second condition.}

\item{fit_type}{String indicating the type of model to fit: "linear" or "cubic"}
}
\value{
A list containing the fit model, fitted values, raw data for displaying, and
the mean, stdev in each condition
}
\description{
Fit a model to the data in each condition.
}

