% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getPvals}
\alias{getPvals}
\title{getPvals.}
\usage{
getPvals(model_stn, iter, model_stn_dist)
}
\arguments{
\item{model_stn}{Vector containing model-based STN for each protein}

\item{iter}{Number of resampling iterations}

\item{model_stn_dist}{Vector containing resampled STN distribution}
}
\value{
Vector containing p-value for each protein
}
\description{
Calculate p-values using resampled STN distribution.
}

