% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSTNdistrib}
\alias{getSTNdistrib}
\title{getSTNdistrib.}
\usage{
getSTNdistrib(A, nA, nB, iter, xbar0_replace, best_model)
}
\arguments{
\item{A}{Data matrix for the baseline (i.e. best fit) condition}

\item{nA}{Number of replicates in the baseline condition}

\item{nB}{Number of replicates in the other condition}

\item{iter}{Number of resampling iterations}

\item{xbar0_replace}{Value to substitute in place of xbar zeros}

\item{best_model}{Best-fit model}
}
\value{
Vector containing model-based STN distribution
}
\description{
Generate STN distribution via resampling.
}

