% Generated by roxygen2 (4.0.2): do not edit by hand
\name{model_fit_plots}
\alias{model_fit_plots}
\title{model_fit_plots.}
\usage{
model_fit_plots(m, outfile)
}
\arguments{
\item{m}{List of model-fit data generated by \code{\link{fit_model}}}

\item{outfile}{Name of output png file containing the plots}
}
\value{
Nothing, prints plots to specified png file
}
\description{
Generate plots showing raw data and model-fit in each condition.
}

