% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stn_pval_plots}
\alias{stn_pval_plots}
\title{stn_pval_plots.}
\usage{
stn_pval_plots(stn_pval, outfile)
}
\arguments{
\item{stn_pval}{List of model-based STN and p-values, as generated by \code{\link{calc_stn_pval}}}

\item{outfile}{Name of PNG file to which plots will be saved}
}
\value{
Nothing, but saves plots to specified output file
}
\description{
Plot STN and p-values for all proteins.
}

